update tests according to the new behavior
