+++ src/tests/conftest.py
@@ -1,7 +1,5 @@
-import sqlite3
-from unittest.mock import MagicMock, patch
+from unittest.mock import patch
 
-import httpx
 import pytest
 
 
@@ -33,17 +31,3 @@ def mock_json_load():
 def mock_json_dump():
     with patch("json.dump", side_effect=json_dump) as mock_json:
         yield mock_json
-
-
-@pytest.fixture()
-def mock_httpx_client_response():
-    with patch.object(httpx.AsyncClient, "request") as mock_response:
-        yield mock_response
-
-
-@pytest.fixture
-def mock_connection():
-    connection = MagicMock(spec=sqlite3.Connection)
-    connection.execute = MagicMock()
-    connection.commit = MagicMock()
-    return connection
